refactor: move array handling pipes to utils file
